test(e2e): update specs for Uma's pre-enabled 2FA + new challenge title
Two casualties of the 2FA changes in this branch:

- `two-factor.spec.js` looked for "Two-factor verification" — the
  challenge step's title is now "Two-factor authentication" to match
  the design mockup.
- `auth.spec.js`'s "fixture standard user denied admin access" test
  expected a password-only sign-in to land on /account. Uma now has
  TOTP 2FA pre-enabled in fixtures so that path stops at the challenge
  step; updated to use a new `signInAsFixtureUser()` helper that walks
  the password + computed-TOTP flow.

Promoted the TOTP-compute helpers from `two-factor.spec.js` into
`helpers.js` (`computeTotpCode`, `FIXTURE_TOTP_SECRET`) since both
specs need them now, and added `signInAsFixtureUser()` as the
canonical "log Uma in through both factors" entry point so future
specs that need a logged-in fixture user don't each redo the dance.
